Fixed k8 tests after refactor
